add Random Walk on debug data.

+++ app/src/main/java/jp/co/mti/marun/android/stargazer/example/NavigationDisplayView.java
@@ -132,7 +132,7 @@ public class NavigationDisplayView extends SurfaceView implements SurfaceHolder.
         mHolder.unlockCanvasAndPost(canvas);
     }
     
-    private int convertMeter2Pixel(float x) {
+    private int convertMeter2Pixel(double x) {
         int p = (int) (GRID_UNIT_PIXEL * x / GRID_UNIT_METER);
         return p;
     }
